add documentation

+++ library/Division.php
@@ -1,26 +1,30 @@
 <?php
 
 /**
- * Created by PhpStorm.
- * User: carlos
- * Date: 25/04/2016
- * Time: 20:40
+ * Class Division
  */
-
 class Division extends OperatorController implements Operation
 {
+    /**
+     * Division constructor.
+     * @param $firstOperator
+     * @param $secondOperator
+     */
     public function __construct($firstOperator, $secondOperator)
     {
        parent::__construct($firstOperator, $secondOperator);
         return true;
     }
 
+    /**
+     * @return float
+     */
     public function doOperation()
     {
         $firstOperator = $this->getFirstOperator();
         $secondOperator = $this->getSecondOperator();
 
-        $result = $firstOperator / $secondOperator;
+        $result = (float)($firstOperator / $secondOperator);
         return $result;
     }
 }
